fix: make decrypt_or_passthrough resilient to decryption failures
Log warnings and return raw ciphertext instead of raising ValueError
when decryption fails (wrong key, missing key, corrupted data). This
prevents a single undecryptable document from crashing entire Beanie
queries.

Also consolidates the log_sink test fixture into conftest.py.

Closes #1564
